requirements.txt
flask>=2.3.0
gunicorn>=20.1.0
dill>=0.4.0
numpy>=1.24.0
pillow>=10.0.0
flask-cors>=4.0.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
opencv-python-headless>=4.8.0
ultralytics>=8.0.0 --no-deps
facenet-pytorch>=2.6.0 